requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters==1.1.1
langchain-chroma==1.1.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langchain-huggingface==1.2.1

langgraph>=0.2.0

langsmith

openai==2.28.0

chromadb>=0.5.0
sentence-transformers>=3.0.0

tavily-python>=0.5.0

PyMuPDF
pymupdf4llm==1.27.2.1

pydantic>=2.0.0
requests>=2.32.0